Create README.md
I am in the early stages of trying to get into cybersecurity. My first stop on this journey is gaining an overall understanding of what is in the CompTIA A+, Network+ and Security+ exams. I have an A+ certification from over ten years ago, so I need to update my knowledge. Next, I will learn Python because cybersecurity and coding go hand in hand. Python is ubiquitous and necessary, so that seems like a good starting point. I want to eventually have a GitHub profile that is attractive to cybersecurity companies and employers to post on LinkedIn. I want to share my experience if I am successful so others can do what I've done to get into cybersecurity.